requirements.txt
Django>=3.0,<4.0
djangorestframework==3.14.0
djoser==2.2.2
django-cors-headers==4.3.1